chore: review comments + rename
