+++ renku/service/controllers/utils.py
@@ -15,7 +15,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Renku service parent remote controller."""
+"""Utilities for renku service controllers."""
 
 import tempfile
 from contextlib import contextmanager
